# Handle views in combo with Ace base templates